fix(element): change folder name to match artifact
During the renaming of Riot to Element, the path for the webapp changed
from /opt/riot to /opt/element. In those folders, the archives are
unpacked, resulting in paths like /opt/element/riot-v${version}. Since
the archives are not yet renamed
(https://github.com/vector-im/element-web/issues/14896), this riot
reference has to stay for now. In the renaming, this was accidentally
changed prematurely, resulting in this regression.