feat(modules): add matrix_member ansible module
This module can manage matrix membership in a given room by inviting, kicking or
banning a list specified users.

With the exclusive=True flag, it can be used to ensure that a given list of
members is in a room (and no one else). For this module, users invited into a
room count as members, as they have permissions to join the room.

Co-authored-by: Jan Christian Grünhage <jan.christian@gruenhage.xyz>