feat: allow setting custom groups for appservices
So far, we hardcoded a `matrix-appservices` group for application
services, which was used so that synapse can read registration files
with read permissions for that group. There's two problems with that:
First, the name was longer than permitted by shadow, so distros that
don't patch out these limitations (like Void Linux) broke when trying to
create that group name. We didn't notice because our production infra
runs on Debian which does patch this out. Secondly though, this enabled
appservices reading the registration files of *other* application
services, which is a potential security risk. Both of these issues can
be resolved with this new feature.