fix(modules/matrix): fix regression in access token handling
52785ab768b30d179b6dc201dada70f059dd4fb7 refactored the login logic,
which, switching to nio's token login when being passed a token. The
token being passed is an *access* token though, not a *login* token, so
that fails and thereby breaks any tasks that use access tokens instead
of username/password for authenticating to matrix.

It also accidentally negated the logic on when to invalidate a token
upon logout, so that has been resolved as well.